🎨 MainPage에서 SidePanel 분리
