+++ src/features/side-panel/index.ts
@@ -1,0 +1,3 @@
+import SidePanel from './SidePanel';
+
+export default SidePanel;
