+++ src/pages/main/MainPage.tsx
@@ -1,9 +1,6 @@
-import LeftArrowIcon from '@/assets/icons/leftArrow.svg?react';
-import ModifyFilledIcon from '@/assets/icons/modifyFilled.svg?react';
 import GlobalNavigationBar from '@/components/global-navigation-bar';
-import CourseView from '@/features/course/course-view';
 import FloatMenu from '@/features/map/float-menu';
-import SearchBar from '@/features/search/search-bar';
+import SidePanel from '@/features/side-panel';
 import { useTmap } from '@/hooks/useTmap';
 
 import * as S from './MainPage.css';
@@ -14,24 +11,9 @@ function MainPage() {
     return (
         <>
             <GlobalNavigationBar />
+
             <div className={S.wrapper}>
-                <div className={S.sidePanel}>
-                    <div className={S.header}>
-                        <LeftArrowIcon
-                            width={24}
-                            height={24}
-                            className={S.backArrowIcon}
-                        />
-                        <p className={S.courseName}>테스트 코스</p>
-                        <ModifyFilledIcon
-                            width={24}
-                            height={24}
-                            className={S.modifyIcon}
-                        />
-                    </div>
-                    <SearchBar />
-                    <CourseView />
-                </div>
+                <SidePanel />
                 <div className={S.map} ref={mapContainerRef}>
                     <FloatMenu />
                 </div>
